Move application config into config folder
We've started to amend the folder structure to split application
configuration up a little.

This change moves the app config into the config folder and out of
the src folder.
