+++ src/controllers/pingdom.controller.js
@@ -1,5 +1,5 @@
 const axios = require('axios')
-const config = require('../config')
+const config = require('../../config')
 
 function returnError (res, error) {
   res.set('Content-Type', 'text/xml')
