+++ src/middleware/user.js
@@ -1,4 +1,4 @@
-const config = require('../config')
+const config = require('../../config')
 const authorisedRequest = require('../lib/authorised-request')
 
 module.exports = (req, res, next) => {
