+++ src/repos/company.repo.js
@@ -1,6 +1,6 @@
 /* eslint camelcase: 0, prefer-promise-reject-errors: 0 */
 const Q = require('q')
-const config = require('../config')
+const config = require('../../config')
 const authorisedRequest = require('../lib/authorised-request')
 
 // Get a company and then pad out the interactions with related data
