+++ src/repos/contact.repo.js
@@ -2,7 +2,7 @@ const winston = require('winston')
 const Q = require('q')
 
 const authorisedRequest = require('../lib/authorised-request')
-const config = require('../config')
+const config = require('../../config')
 
 function getContact (token, contactId) {
   return authorisedRequest(token, `${config.apiRoot}/v3/contact/${contactId}`)
