+++ src/repos/metadata.repo.js
@@ -1,6 +1,6 @@
 const winston = require('winston')
 const authorisedRequest = require('../lib/authorised-request')
-const config = require('../config')
+const config = require('../../config')
 
 let redisClient
