+++ src/services/search.service.js
@@ -1,5 +1,5 @@
 const authorisedRequest = require('../lib/authorised-request')
-const config = require('../config')
+const config = require('../../config')
 const { buildQueryString } = require('../lib/url-helpers')
 
 function search ({ token, searchTerm, searchType, limit = 10, page = 1 }) {
