+++ test/controllers/search.controller.test.js
@@ -1,5 +1,5 @@
 const nock = require('nock')
-const config = require('~/src/config')
+const config = require('~/config')
 
 const next = function (error) {
   throw Error(error)
